+++ src-core/modules/network/module_network_client.cpp
@@ -52,14 +52,14 @@ namespace network
 
         nng_sub0_open_raw(&sock);
         nng_dialer_create(&dialer, sock, std::string("tcp://" + address + ":" + std::to_string(port)).c_str());
-        nng_dialer_start(dialer, NULL);
+        nng_dialer_start(dialer, (int)NULL);
 
         while (input_active.load())
         {
             size_t lpkt_size;
-            nng_recv(sock, buffer, &lpkt_size, NULL);
+            nng_recv(sock, buffer, &lpkt_size, (int)NULL);
 
-            if (pkt_size != lpkt_size)
+            if (pkt_size != (int)lpkt_size)
                 continue;
 
             output_fifo->write(buffer, pkt_size);
